Remove the call to `checkStorage`
This is really the responsibility of either the code configuring the storage (if it can assume
the storage stays good for the life-time of the program), or it must be part of the storage
itself (if that storage is so flaky that it might fail any moment).

Running a potentially expensive write-then-read cycle while working on setting up a TLS
connection is not ok.